[refactor] test page로 분리
