+++ src/pages/test/TestBottomSheetTutorialPage.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { BOTTOM_POSITION, BottomSheetLayout, MIDDLE_POSITION, TOP_POSITION } from "../layouts/BottomSheetLayout";
+import { BOTTOM_POSITION, BottomSheetLayout, MIDDLE_POSITION, TOP_POSITION } from "../../layouts/BottomSheetLayout";
 
 export const TEST_BOTTOM_SHEET_PAGE_PATH = "/test/bottom/sheet/diary";
